+++ axum-login/src/user_store.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 
-use crate::{AuthUser, Result};
+use crate::AuthUser;
 
 /// A trait which defines a method that allows retrieval of users from an
 /// arbitrary backend.
@@ -12,6 +12,9 @@ where
     /// An associated user type which will be loaded from the store.
     type User: AuthUser<UserId, Role>;
 
+    /// Associated error type
+    type Error: std::error::Error + Send;
+
     /// Load and return a user.
     ///
     /// This provides a generic interface for loading a user from some store.
@@ -19,5 +22,5 @@ where
     /// unique, stable identifier of the user is available. See [`AuthUser`]
     /// for expected minimal interface of the user type itself.
     #[must_use]
-    async fn load_user(&self, user_id: &UserId) -> Result<Option<Self::User>>;
+    async fn load_user(&self, user_id: &UserId) -> Result<Option<Self::User>, Self::Error>;
 }